move slip reader out of web serial transport
